Add hot loader
