+++ .babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["latest", "react"]
+  "presets": ["latest", "react"],
+  "plugins": ["react-hot-loader/babel"]
 }
