+++ initializers/server/index.js
@@ -4,9 +4,10 @@ const webpackDevServer = require('webpack-dev-server');
 const config = require('../../webpack.config.js');
 
 const host = 'localhost';
-const port = 3001;
+const port = 3000;
 
 new webpackDevServer(webpack(config), {
+  hot: true,
   publicPath: config.output.publicPath,
   stats: { color: true }
 }).listen(port, host, (err) => {
