+++ package.json
@@ -26,6 +26,7 @@
     "eslint-plugin-destructuring": "^2.1.0",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-react": "^7.4.0",
+    "react-hot-loader": "^3.0.0",
     "webpack": "^3.8.1",
     "webpack-dev-server": "^2.9.4"
   },
